Moved operator mailer stuff out of TemplateBasedMailer
As per discussion on https://github.com/wmde/FundraisingFrontend/pull/679
